requirements.txt
streamlit==1.33.0
pymongo==4.6.1
chromadb==0.4.24
langchain==0.1.16
langchain-groq==0.1.4
python-dotenv==1.0.1
requests==2.31.0
plotly==5.20.0
google-generativeai==0.5.2
beautifulsoup4==4.12.3
protobuf==4.25.3
